Implemented Login Functionality.
Users can now log in. They are redirected to the home page where they
are greeted with a flash message indicated a successfully log in. If
they fail to log in due to invalid credentials, they are redirected to
the login page and greeted with a red flash messages indicated failure.