Reverting back to common js syntax from es module syntax.
Jaycee wasn't able to run the app on her machine. It was causing a
segmentation fault. A TA at office hours suggested that it may be because es modules
is a new experimental feature. Going back to require means giving up a
library that we wanted to use, but the most important thing is that
everyone can work on the project. So, I created this branch to see if
going back to 'require' instead of 'import' fixes the issue for my
teammate.