Pair programming with Jason. Add card.
Together, we implemented the ability to search for Pokemon by name and
add them to the users collection by clicking a button. If user is not
logged in, button to add card to collection does not show. But the
user can still search Pokemon info. The user may have multiple of the
same card.

If the searched card is not found, a message will indicate that the
card couldn't be found. Otherwise, the card is added to the users
collection and a flash message indicates success.

Changed the database to accommodate multiple of the same card for one
user. Previously there was a unique constraint on the users_pokemon table.