Trade Page Created.
The purpose of this page is to find a partner to
trade cards with. In this page, the user can see, on a right pane, the other users on
the sight. They may select one of them, and they will see the users
cards on the left.

They can click a button to be brought to a page
where they can specify the cards they want to trade and initiate that
trade, but that page is unimplemented as of this commit.