Browse Trade Partners Page Implemented, Request Trade Page also
In progress on defining the backend logic to handle a requested trade
with another user. At this point, you can view other users collections
and select another user to trade with. This brings you to another page
where you can select which cards you would like to give and receive.