Pending Trades Inbox, Accept and Decline Trades
Now, there is page to view your incoming and outgoing trade
requests. You can view these trades by clicking the links. For any
trade in which you are involved, you can decline the trade. And for
trades which you did not initiate, you can accept the trade. The
changes will be reflected in your collection and your trade partners
collection.

If you are involved in multiple pending trades, and you accept a trade
which invalidates some other trades, then those trades will still
appear for now, but they will not be committed to the database or
cause errors if you attempt to accept them.